Fixed project architecture. Added create book endpoint.

+++ internal/service/user.go
@@ -16,20 +16,20 @@ func NewUserService(repo repository.UserRepoInterface) *UserService {
 	}
 }
 
-func (us *UserService) Create(ctx context.Context, user model.User) (int, error) {
+func (us *UserService) Create(ctx context.Context, user model.User) (uint64, error) {
 	return 0, nil
 }
 func (us *UserService) GetAll(ctx context.Context) ([]model.User, error) {
 	var users []model.User
 	return users, nil
 }
-func (us *UserService) GetById(ctx context.Context, userId int) (model.User, error) {
+func (us *UserService) GetById(ctx context.Context, userId uint64) (model.User, error) {
 	var user model.User
 	return user, nil
 }
-func (us *UserService) Delete(ctx context.Context, userId int) error {
+func (us *UserService) Delete(ctx context.Context, userId uint64) error {
 	return nil
 }
-func (us *UserService) Update(ctx context.Context, userId int) error {
+func (us *UserService) Update(ctx context.Context, userId uint64) error {
 	return nil
 }
